# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Photo Viewer API",
        "contact": {},
        "version": "dev"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/collections": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "List collections",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Owner login; without the param — all visible collections",
                        "name": "owner_login",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/service.SimpleCollectionInfo"
                            }
                        }
                    },
                    "404": {
                        "description": "owner not found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "Create a collection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JSON: title (≤50), description (≤200), access_level (private|protected|public), photo_uuids []",
                        "name": "metadata",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/upload.CollecionCreateResponse"
                        }
                    },
                    "400": {
                        "description": "invalid metadata / validation",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "token is empty / invalid token",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "quota is not enough / collection already exists",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "internal error (incl. unknown photo in photo_uuids)",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/tags": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "List tags",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter tags by photo UUID",
                        "name": "photo_uuid",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/service.TagInfo"
                            }
                        }
                    },
                    "400": {
                        "description": "invalid photo uuid / photo not found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Create a tag",
                "parameters": [
                    {
                        "description": "tag_name (2-50, unique), tag_description (≤200)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.SaveTagInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/upload.TagCreateResponse"
                        }
                    },
                    "400": {
                        "description": "invalid metadata / validation / tag with this name already exists",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "token is empty / invalid token",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/users/{user_uuid}": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update user (role / quotas)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User UUID",
                        "name": "user_uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "user_role (user|moderator), photos_quota (≥0), collections_quota (≥0)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.PatchUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "invalid metadata / validation / invalid request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "token is empty / invalid token",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "access denied (admin role required)",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "user not found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "entity.AccessModifier": {
            "type": "string",
            "enum": [
                "private",
                "protected",
                "public"
            ],
            "x-enum-varnames": [
                "AccessModifierPrivate",
                "AccessModifierProtected",
                "AccessModifierPublic"
            ]
        },
        "entity.UserRole": {
            "type": "string",
            "enum": [
                "user",
                "moderator",
                "admin"
            ],
            "x-enum-varnames": [
                "UserDefaultRole",
                "UserModeratorRole",
                "UserAdminRole"
            ]
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Some error description"
                },
                "status": {
                    "type": "string",
                    "example": "Ok"
                }
            }
        },
...
        "service.PatchUserRequest": {
            "type": "object",
            "properties": {
                "collections_quota": {
                    "type": "integer",
                    "minimum": 0
                },
                "photos_quota": {
                    "type": "integer",
                    "minimum": 0
                },
                "user_role": {
                    "enum": [
                        "user",
                        "moderator"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.UserRole"
                        }
                    ]
                }
            }
        },
...
        "service.SaveTagInput": {
            "type": "object",
            "required": [
                "tag_name"
            ],
            "properties": {
                "tag_description": {
                    "type": "string",
                    "maxLength": 200
                },
                "tag_name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2
                }
            }
        },
        "service.SimpleCollectionInfo": {
            "type": "object",
            "required": [
                "access_level"
            ],
            "properties": {
                "access_level": {
                    "$ref": "#/definitions/entity.AccessModifier"
                },
                "collection_uuid": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "maxLength": 200
                },
                "owner_login": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "maxLength": 50
                }
            }
        },
        "service.TagInfo": {
            "type": "object",
            "properties": {
                "tag_description": {
                    "type": "string"
                },
                "tag_name": {
                    "type": "string"
                },
                "tag_uuid": {
                    "type": "string"
                }
            }
        },
...
        "upload.CollecionCreateResponse": {
            "type": "object",
            "properties": {
                "collection_uuid": {
                    "type": "string"
                },
                "error": {
                    "type": "string",
                    "example": "Some error description"
                },
                "status": {
                    "type": "string",
                    "example": "Ok"
                }
            }
        },
...
        "upload.TagCreateResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Some error description"
                },
                "status": {
                    "type": "string",
                    "example": "Ok"
                },
                "tag_uuid": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Format: \"Bearer \u003caccess_token\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}